financial-rag-chatbot/requirements.txt
pandas
requests
openai
langchain
streamlit # or another framework for your UI
alpha_vantage
python-dotenv